feat: add internationalized rss feed
